interactor run trait file rename
