+++ framework/snippets/src/interactor/interactor_tx.rs
@@ -6,17 +6,17 @@ mod interactor_exec_env;
 mod interactor_exec_step;
 mod interactor_exec_transf;
 mod interactor_exec_upgrade;
-mod interactor_prepare_async;
 mod interactor_query_call;
 mod interactor_query_env;
 mod interactor_query_step;
+mod interactor_run_trait;
 mod simulate_gas_marker;
 
 pub use interactor_exec_env::InteractorEnvExec;
 pub use interactor_exec_step::InteractorExecStep;
-pub use interactor_prepare_async::{
-    InteractorPrepareAsync, InteractorRunAsync, InteractorSimulateGasAsync,
-};
 pub use interactor_query_env::InteractorEnvQuery;
 pub use interactor_query_step::InteractorQueryStep;
+pub use interactor_run_trait::{
+    InteractorPrepareAsync, InteractorRunAsync, InteractorSimulateGasAsync,
+};
 pub use simulate_gas_marker::SimulateGas;
